update nnunet format
