+++ medicalseg/models/layers/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .layer_libs import Identity
+from .layer_libs import Identity, ConvDropoutNormNonlin
